Move oneifs to separate package.
