+++ cmd/onefs/rest.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD style
 // license that can be found in the LICENSE file.
 
-package iquota
+package onefs
 
 type QuotaParams struct {
 	User  string `schema:"user"`
@@ -14,15 +14,4 @@ type QuotaParams struct {
 type QuotaRestResponse struct {
 	Default *Quota   `json:"default"`
 	Quotas  []*Quota `json:"quotas"`
-}
-
-type IQuota struct {
-	Path            string `json:"path"`
-	GracePeriod     string `json:"pretty_grace_period"`
-	HardLimit       int    `json:"hard_limit"`
-	SoftLimit       int    `json:"soft_limit"`
-	Used            int    `json:"used"`
-	HardLimitInodes int    `json:"hard_limit_inodes"`
-	SoftLimitInodes int    `json:"soft_limit_inodes"`
-	UsedInodes      int    `json:"used_inodes"`
 }
